requirements.txt
cryptography>=3.4.8
Pillow>=9.0.0
pytest>=7.0.0
numpy>=1.20.0
